refactor: 일부 수정
