+++ src/components/CardList/CardOverview.module.scss
@@ -81,7 +81,7 @@
   }
 }
 
-.orange {
+.beige {
   background-color: $color-orange-200;
 }
 .purple {
