+++ src/components/CardList/Carousel.module.scss
@@ -3,7 +3,7 @@
   flex-direction: column;
   gap: 16px;
   overflow: hidden;
-  padding: 24px;
+  padding: 0 24px;
 
   & .title {
     color: $color-black;
