+++ src/pages/listpage/ListPage.jsx
@@ -1,17 +1,16 @@
-import CardList from '../../components/CardList/CardList';
+import Carousel from '../../components/CardList/Carousel';
 import styles from './ListPage.module.scss';
 import useFetch from '../../hooks/useFetch';
-import Button from '../../components/Button/Button/Button';
+import { Button } from '../../components/Button/';
 import LoadingPage from '../LoadingPage/LoadingPage';
+import CardList from '../../components/CardList/CardList';
 
 // 안된것들
 // 로딩페이지 미구현 - 채민님
-// CardList 애니매이션 미구현 behavior: 'smooth',
-// 애니매이션은 애플공홈 참고
+// CardList 애니매이션 - carousel ?
 // 버튼 미구현
 // 반응형 미구현
 // 무한스크롤? 가능함?
-// ArrowBtn svg 작업?
 
 export default function ListPage() {
   const { data, isLoading } = useFetch(`/2-7/recipients/`);
@@ -32,16 +31,19 @@ export default function ListPage() {
     <main className={styles.container}>
       <>
         <div>
-          <CardList
+          <Carousel
             CardListName={'인기 롤링 페이퍼 🔥'}
             recipients={hotRecipients}
           />
         </div>
         <div>
-          <CardList
+          <Carousel
             CardListName={'최근에 만든 롤링 페이퍼 💜'}
             recipients={newRecipients}
           />
+        </div>
+        <div>
+          <CardList CardListName={`전체 롤링페이퍼`} recipients={recipients} />
         </div>
         <Button className={styles.myButton} size={'md'}>
           나도 만들어보기
